Move enums to a separate header with inline ostream operator
